pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icr"
version = "1.0.0"
description = "Infinite Context Runtime - RLM-inspired unlimited context for Claude Code (Research Grade)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "ICR Contributors" }
]
keywords = [
    "code-search",
    "semantic-search",
    "embeddings",
    "claude",
    "mcp",
    "code-intelligence",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Indexing",
]

dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "numpy>=1.24",
    "hnswlib>=0.7",
    "sentence-transformers>=2.2",
    "watchdog>=3.0",
    "tree-sitter>=0.20",
    "tree-sitter-languages>=1.8",
    "mcp>=1.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
    "pre-commit>=3.0",
]

full = [
    "icd>=0.1.0",
    "ic-mcp>=0.1.0",
    "ic-claude>=0.1.0",
]

[project.scripts]
icr = "icr.cli:main"

[project.urls]
Homepage = "https://github.com/anthropics/icr"
Documentation = "https://github.com/anthropics/icr#readme"
Repository = "https://github.com/anthropics/icr"
Issues = "https://github.com/anthropics/icr/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/icr"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "LICENSE",
    "README.md",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict= (Python 3.10+)
]

[tool.ruff.lint.isort]
known-first-party = ["icr"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "hnswlib",
    "sentence_transformers",
    "tree_sitter",
    "tree_sitter_languages",
    "mcp",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/icr"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
